requirements.txt
azure-cognitiveservices-speech==1.38.0
keyboard==0.13.5
openai==1.41.1